+++ _/app/index.js
@@ -5,7 +5,13 @@ document.body.append(span);
 const worker = new Worker('./web-worker.bundle.js');
 
 worker.addEventListener('message', ({data}) => {
-    console.log('frontend: message received: ', data);
+    switch (data.msg) {
+        case 'currencies':
+            console.table(data.payload);
+            break;
+        default:
+            console.log('frontend: message received: ', data)
+    }
 })
 
 worker.postMessage('hello, worker');