+++ _/worker/API/api.js
@@ -1,0 +1,3 @@
+module.exports = {
+    CURRENCIES: 'https://www.nbrb.by/api/exrates/currencies',
+}