+++ _/worker/API/getCurrencies.js
@@ -1,0 +1,5 @@
+const fetchJSON = require('../utils/fetchJSON');
+const { CURRENCIES } = require('./api');
+const currenciesMapper = require('../mappers/currencies');
+
+module.exports = () => fetchJSON(CURRENCIES).then(currenciesMapper);