+++ _/worker/index.js
@@ -1,3 +1,9 @@
+const getCurrencies = require('./API/getCurrencies');
+
+getCurrencies()
+    .then(payload => ({ msg: 'currencies', payload}))
+    .then(postMessage);
+
 addEventListener('message', ({data}) => {
     console.log('worker: message received', data);