+++ _/worker/mappers/currencies.js
@@ -1,0 +1,3 @@
+const currencyMapper = require('./currency');
+
+module.exports = (currencies) => currencies.map(currencyMapper);