+++ _/worker/mappers/currency.js
@@ -1,0 +1,10 @@
+module.exports = ({
+    Cur_ID: id,
+    Cur_Code: code,
+    Cur_Abbreviation: abbr,
+    Cur_Name: name,
+    Cur_Name_Eng: nameEng,
+    Cur_Scale: scale,
+}) => ({
+    id, code, abbr, name, nameEng, scale
+})