+++ _/worker/utils/fetchJSON.js
@@ -1,0 +1,4 @@
+module.exports = function(url) {
+    return fetch(url)
+        .then(response => response.json())
+}